Remove net/http transport wrappers, add new Client methods.
Remove the net/http transport wrappers because the wrappers are
inefficient and not helpful.

Add Post and Get methods. These methods sign and invoke a POST and GET
request respectively.

Add SignFrom method. SignForm is nearly identical to the now deprecated
SignParam. The difference is that SignForm panics if the application
erroneously includes a query string in the URL.

Improve package-level documentation.