Improve auth header construction.
The code to add comma separators to the authorization header value
relied on the fact that the first element in oauthKeys is always present
in the header value. Make the code more robust by checking if a
parameter has already been added to the header before adding the comma
separator.